third_party/blink/web_tests/platform/generic/gamepad/full-screen-gamepad-expected.txt

EVENT(gamepadconnected)
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLHtmlElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLSpanElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLHtmlElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == 'undefined') OK
END OF TEST
